control: extract _ThreadedCommandRunner to a separate module
