+++ lib/control/control-server.c
@@ -23,103 +23,8 @@
 
 #include "control-server.h"
 #include "control-connection.h"
-#include "mainloop.h"
+#include "control-command-thread.h"
 #include "messages.h"
-
-typedef struct _ThreadedCommandRunner
-{
-  ControlConnection *connection;
-  GString *command;
-  gpointer user_data;
-
-  GThread *thread;
-  struct
-  {
-    GMutex state_lock;
-    gboolean cancelled;
-    gboolean finished;
-  } real_thread;
-  ControlCommandFunc func;
-  struct iv_event thread_finished;
-} ThreadedCommandRunner;
-
-static ThreadedCommandRunner *
-_thread_command_runner_new(ControlConnection *cc, GString *cmd, gpointer user_data)
-{
-  ThreadedCommandRunner *self = g_new0(ThreadedCommandRunner, 1);
-  self->connection = control_connection_ref(cc);
-  self->command = g_string_new(cmd->str);
-  self->user_data = user_data;
-
-  g_mutex_init(&self->real_thread.state_lock);
-
-  return self;
-}
-
-static void
-_thread_command_runner_free(ThreadedCommandRunner *self)
-{
-  g_mutex_clear(&self->real_thread.state_lock);
-  g_string_free(self->command, TRUE);
-  control_connection_unref(self->connection);
-  g_free(self);
-}
-
-static void
-_on_thread_finished(gpointer user_data)
-{
-  ThreadedCommandRunner *self = (ThreadedCommandRunner *) user_data;
-  g_thread_join(self->thread);
-  iv_event_unregister(&self->thread_finished);
-  ControlServer *server = self->connection->server;
-  server->worker_threads = g_list_remove(server->worker_threads, self);
-  _thread_command_runner_free(self);
-}
-
-static void
-_thread(gpointer user_data)
-{
-  ThreadedCommandRunner *self = (ThreadedCommandRunner *)user_data;
-  self->func(self->connection, self->command, self->user_data);
-  g_mutex_lock(&self->real_thread.state_lock);
-  {
-    self->real_thread.finished = TRUE;
-    if (!self->real_thread.cancelled)
-      iv_event_post(&self->thread_finished);
-  }
-  g_mutex_unlock(&self->real_thread.state_lock);
-}
-
-static void
-_thread_command_runner_sync_run(ThreadedCommandRunner *self, ControlCommandFunc func)
-{
-  msg_warning("Cannot start a separated thread - ControlServer is not running",
-              evt_tag_str("command", self->command->str));
-
-  func(self->connection, self->command, self->user_data);
-  _thread_command_runner_free(self);
-}
-
-static void
-_thread_command_runner_run(ThreadedCommandRunner *self, ControlCommandFunc func)
-{
-  IV_EVENT_INIT(&self->thread_finished);
-  self->thread_finished.handler = _on_thread_finished;
-  self->thread_finished.cookie = self;
-  self->func = func;
-
-  if (!main_loop_is_control_server_running(main_loop_get_instance()))
-    {
-      _thread_command_runner_sync_run(self, func);
-      return;
-    }
-
-  iv_event_register(&self->thread_finished);
-
-  self->thread = g_thread_new(self->command->str, (GThreadFunc) _thread, self);
-  ControlServer *server = self->connection->server;
-  server->worker_threads = g_list_append(server->worker_threads, self);
-}
 
 void
 control_connection_start_as_thread(ControlConnection *self, ControlCommandFunc cmd_cb,
@@ -127,29 +32,6 @@ control_connection_start_as_thread(ControlConnection *self, ControlCommandFunc c
 {
   ThreadedCommandRunner *runner = _thread_command_runner_new(self, command, user_data);
   _thread_command_runner_run(runner, cmd_cb);
-}
-
-static void
-_delete_thread_command_runner(gpointer data)
-{
-  ThreadedCommandRunner *self = (ThreadedCommandRunner *) data;
-  gboolean has_to_free = FALSE;
-
-  g_mutex_lock(&self->real_thread.state_lock);
-  {
-    self->real_thread.cancelled = TRUE;
-    if (!self->real_thread.finished)
-      {
-        has_to_free = TRUE;
-      }
-  }
-  g_mutex_unlock(&self->real_thread.state_lock);
-
-  if (has_to_free)
-    {
-      g_thread_join(self->thread);
-      _thread_command_runner_free(self);
-    }
 }
 
 void
